api/pyproject.toml
[project]
name = "manhattan-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.12",
    "beanie>=1.30.0",
    "fastapi[standard]>=0.115.9",
    "google-genai>=1.19.0",
    "google-generativeai>=0.8.5",
    "motor>=3.7.1",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
]

[tool.black]
line-length = 160
